added time view on task bar

+++ core/assets/os/components/TablePositionComponent.groovy
@@ -3,5 +3,5 @@ package os.components;
 import com.badlogic.ashley.core.*;
 
 public class TablePositionComponent implements Component {
-	public int x, y;
+	public float x, y;
 }
